Changed ce_pin and csn_pin to uint16_t so the library supports GPIO identifiers higher than 255 (Minnowboard)

+++ RF24.h
@@ -64,8 +64,8 @@ private:
   GPIO gpio;
 #endif
 
-  uint8_t ce_pin; /**< "Chip Enable" pin, activates the RX or TX role */
-  uint8_t csn_pin; /**< SPI Chip select */
+  uint16_t ce_pin; /**< "Chip Enable" pin, activates the RX or TX role */
+  uint16_t csn_pin; /**< SPI Chip select */
   uint16_t spi_speed; /**< SPI Bus Speed */
 #if defined (RF24_LINUX)
   uint8_t spi_rxbuff[32+1] ; //SPI receive buffer (payload max 32 bytes)
@@ -108,7 +108,7 @@ public:
    * @param _cepin The pin attached to Chip Enable on the RF module
    * @param _cspin The pin attached to Chip Select
    */
-  RF24(uint8_t _cepin, uint8_t _cspin);
+  RF24(uint16_t _cepin, uint16_t _cspin);
   //#if defined (RF24_LINUX)
   
     /**
@@ -122,7 +122,7 @@ public:
   * @param spispeed For RPi, the SPI speed in MHZ ie: BCM2835_SPI_SPEED_8MHZ
   */
   
-  RF24(uint8_t _cepin, uint8_t _cspin, uint32_t spispeed );
+  RF24(uint16_t _cepin, uint16_t _cspin, uint32_t spispeed );
   //#endif
 
   #if defined (RF24_LINUX)
@@ -1886,4 +1886,3 @@ private:
  */
 
 #endif // __RF24_H__
-
